defined class, Musician and other classes with their methods and attributes
