--- pythonic_garage_band/pythonic_garage_band.py
@@ -1,11 +1,0 @@
-class band:
-    def __init__ (self, name, members):
-        self.name = name
-        self.members = []
-    
-    def play_solos():
-        pass
-
-    def to_list():
-        pass
-    # methods: play_solos +  to_list
